implement test service with exception
