+++ Dockerfile
@@ -5,11 +5,10 @@ FROM python:3.9.0
 RUN mkdir /app/
 WORKDIR /app
 
-COPY src/<<INSERT_FOLDER_NAME>> src/<<INSERT_FOLDER_NAME>>
+COPY src/TestCalculationService src/TestCalculationService
 COPY requirements.txt ./
 COPY pyproject.toml ./
 COPY README.md ./
-RUN pip install -r requirements.txt
-RUN pip install ./
+RUN pip install -r requirements.txt && pip install ./
 
-ENTRYPOINT python3 src/<<INSERT_FOLDER_NAME>>/<<INSERT_MAIN_PYTHON_FILENAME>>.py
+ENTRYPOINT ["python3", "src/TestCalculationService/TestService.py"]
