+++ pyproject.toml
@@ -3,12 +3,12 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "ExampleCalculationService"
+name = "TestCalculationService"
 version = "0.0.1"
 authors = [
   { name="Example Author", email="author@example.com" },
 ]
-description = "A small example package"
+description = "An example calculation service"
 readme = "README.md"
 requires-python = ">=3.8"
 classifiers = [
