+++ requirements.txt
@@ -1,2 +1,2 @@
-helics==3.4.0
-dots_infrastructure==0.3.9
+helics==3.6.1
+dots-infrastructure==0.4.6
